Move logic from bin entrypoints into library crate
